fw-kdevops: fix dhcp races
When booting the latest linux-next I see that network interfaces
are up but dhcp never kicks in. It seems we have yet again a name
change issue... Seriously...

If we are not doing network development we frankly don't care about
this bug, so just allow us to work around it. *We* firmware API
kernel developers don't care about this bug so just work around this
nasty issue by ensuring all of our non loopback devices *get* a
DHCP request issued.

We use a new ansible role for this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>